a little hackish
I would have liked to handle the toggle state internally in the listctrl class however I cant get the listctrl col click event to fire in the super class. So have left it to the listpanel to handle the change toggle state. with the magnet_listpanel on_col_left_click returning the super class method