added a toggle_col boolean
if set to false then left column click event is not bound to the listctrl and toggle wont happen. If set to true then toggle will be bound. This only is valid for the magnet list ctrl as want to sort items depending on column header click. use a mixin listcrl auto width for other listctrls